+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-## 0.5.12 (WIP)
+## 0.6.0 (WIP)
 
 * Make `Pledge` polyfill for `Promise`
 * Add `Pledge.after(n, val)`, which returns a pledge that resolves after n milliseconds
 * Add `Pledge#race(other_pledge)`
 * Add new `Array#findByPath(obj)` signature
 * Add `Object.sizeof(input)` to get the size of a variable in bytes
+* Add `Cache` class
 
 ## 0.5.11 (2018-07-14)